requirements.txt
fastapi
uvicorn
pydantic
pytest
pytest-asyncio
sqlalchemy
psycopg2-binary
passlib[bcrypt]
redis
httpx
python-dotenv
black
flake8
isort
pre-commit
PyJWT
pydantic-settings
